requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.0
groq==0.9.0
qdrant-client==1.9.1
firebase-admin==6.5.0
torch==2.3.0
numpy==1.26.4
scikit-learn==1.5.0
python-multipart==0.0.9
httpx==0.27.0
schedule==1.2.2
pydantic==2.7.0
